[dxvk] Switch to implicit VRR frame pacing
vkWaitForPresentKHR() times are not reliable to determine V-Blank times, as they depend on tight guarantees of the compositor and the graphics driver. The previous VRR mode regularly went into V-Sync buffering and its usage on Wayland was limited. Although we couldn't measure the V-Sync buffering to this fine degree, this became pretty clear during play-testing.

This implicit VRR pacing derives V-Blanks based on render-finished times and completely ignores any results from vkWaitForPresentKHR(). This also greatly improves compatibility and it now can be used on Xwayland and native Wayland.